requirements.txt
firebase_functions~=0.1.0
google-cloud-translate==3.12.0
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
tweepy==4.14.0